feat: pass a new res.locals.params object to http-incoming
This is in line with Fastify podlet and layout which provide reply.app.params to http-incoming.
It is not, however, in line with Podium layout which passes the whole res.locals object. Passing
the whole res.locals object can cause circular reference issues since we set http-incoming on res.locals.podium as well and should be changed.